move select "drop" box calculation to Score component and separate the box into its own component
