+++ src/components/Score.js
@@ -28,6 +28,71 @@ class Score extends Component {
     };
   }
 
+  getSelectedRangeForSingleRow = (measure, xStart, xEnd) => {
+    if(xStart > measure.xOfMeasure && xEnd < measure.xOfMeasure + measure.width) {
+      return {
+        start: xStart - measure.xOfMeasure,
+        width: xEnd - xStart
+      };
+    } else if((xStart < measure.xOfMeasure && xEnd > measure.xOfMeasure) ||
+       (xStart > measure.xOfMeasure && xStart < measure.xOfMeasure + measure.width)) {
+
+      if(measure.xOfMeasure < xStart) {
+        return {
+          start: xStart - measure.xOfMeasure,
+          width: measure.width
+        };
+      } else if(xEnd < measure.xOfMeasure + measure.width) {
+        return {
+          start: 0,
+          width: xEnd - measure.xOfMeasure
+        };
+      } else {
+        return 'all';
+      }
+    } else {
+      return undefined;
+    }
+  };
+
+  getSelectedRange = (measure, xStart, xEnd, selectedRows) => {
+    if(!selectedRows) {
+      return undefined;
+    }
+    const selectedRowIndex = selectedRows.indexOf(measure.rowIndex);
+    if(selectedRowIndex === -1) {
+      return undefined;
+    }
+
+    if(selectedRows.length === 1) {
+      return this.getSelectedRangeForSingleRow(measure, xStart, xEnd);
+    } else if(selectedRowIndex === 0) {
+      if(measure.xOfMeasure + measure.width > xStart) {
+        if(measure.xOfMeasure < xStart) { // starting measure, need note index
+          return {
+            start: xStart - measure.xOfMeasure,
+            width: measure.xOfMeasure + measure.width
+          };
+        } else {
+          return 'all';
+        }
+      }
+    } else if(selectedRowIndex === selectedRows.length - 1) {
+      if(xEnd > measure.xOfMeasure) {
+        if(xEnd < measure.xOfMeasure + measure.width) { // last measure
+          return {
+            start: 0,
+            width: xEnd - measure.xOfMeasure
+          };
+        } else {
+          return 'all';
+        }
+      }
+    } else {
+      return 'all';
+    }
+  };
+
   onMouseDown = (e) => {
     e.preventDefault();
 
@@ -38,9 +103,7 @@ class Score extends Component {
       dragStart: { x, y },
       x,
       y,
-      selectedRows: undefined,
-      xStart: undefined,
-      xEnd: undefined
+      selectedRanges: undefined
     });
   };
 
@@ -48,12 +111,14 @@ class Score extends Component {
     // TODO call action creator to define the select range
     const { x, y, dragHeight, dragWidth } = this.state;
 
-    let startRow, endRow, selectedRows;
+    let selectedRows;
     if(dragWidth > 5 && dragHeight > 5) {
-      startRow = Math.floor(y / this.props.rowHeight);
-      endRow = Math.floor((y + dragHeight) / this.props.rowHeight);
+      const startRow = Math.floor(y / this.props.rowHeight);
+      const endRow = Math.floor((y + dragHeight) / this.props.rowHeight);
       selectedRows = Array.from({ length: endRow - startRow + 1 }, (_, k) => k + startRow);
     }
+
+    const selectRanges = this.props.measures.map(measure => this.getSelectedRange(measure, x, x + dragWidth, selectedRows));
 
     this.setState({
       dragStart: undefined,
@@ -61,9 +126,7 @@ class Score extends Component {
       y: undefined,
       dragWidth: undefined,
       dragHeight: undefined,
-      selectedRows,
-      xStart: x,
-      xEnd: x + dragWidth
+      selectRanges
     });
   };
 
@@ -85,12 +148,12 @@ class Score extends Component {
 
   render() {
     const { height, width, measures } = this.props;
-    const { x, y, dragWidth, dragHeight, selectedRows, xStart, xEnd } = this.state;
+    const { x, y, dragWidth, dragHeight, selectRanges } = this.state;
 
     return (
       <div style={{ ...style, height, width }}
         onMouseDown={this.onMouseDown} onMouseUp={this.onMouseUp} onMouseMove={this.onMouseMove}>
-        { measures.map((_, i) => <Measure key={i} measureIndex={i} xStart={xStart} xEnd={xEnd} selectedRows={selectedRows} />) }
+        { measures.map((_, i) => <Measure key={i} measureIndex={i} selected={selectRanges ? selectRanges[i] : undefined} />) }
         <SelectBox height={height} width={width} x={x} y={y} dragWidth={dragWidth} dragHeight={dragHeight} />
       </div>
     );
